@@ -1,7 +1,6 @@
 [package]
-edition = "2018"
 name = "crev-common"
-version = "0.23.2"
+version = "0.23.5"
 description = "Scalable, social, Code REView system that we desperately need - common code"
 authors = ["Dawid Ciężarkiewicz <dpc@dpc.pw>"]
 documentation = "https://docs.rs/crev-lib"
@@ -10,17 +9,18 @@ homepage = "https://github.com/crev-dev/crev"
 keywords = ["code", "review", "trust", "distributed"]
 license = "MPL-2.0 OR MIT OR Apache-2.0"
 readme = "README.md"
+edition = "2021"
 
 [dependencies]
 base64 = "0.13"
 blake2 = "0.9.2"
-chrono = { version = "0.4.19", default-features = false, features = ["std", "clock"] }
+chrono = { version = "0.4.23", default-features = false, features = ["std", "clock"] }
 hex = "0.4.3"
-rprompt = "1.0.5"
-serde = "1.0.139"
-serde_yaml = "0.9.4"
+rprompt = "2.0.2"
+serde = "1.0.150"
+serde_yaml = "0.9.14"
 shell-escape = "0.1.5"
-thiserror = "1.0.31"
+thiserror = "1.0.37"
 rand = "0.8.5"
 
 [package.metadata.release]
