@@ -2,7 +2,7 @@ use crate::{proof, proof::Proof, Error, ParseError, Result};
 use chrono::{self, prelude::*};
 use crev_common::{
     self,
-    serde::{as_rfc3339_fixed, from_rfc3339_fixed},
+    serde::{as_base64, as_rfc3339_fixed, from_base64, from_rfc3339_fixed},
 };
 use derive_builder::Builder;
 use serde::{self, Deserialize, Serialize};
@@ -54,6 +54,16 @@ pub trait CommonOps {
     }
 }
 
+/// Reference to original proof when reissuing
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct OriginalReference {
+    /// original proof digest (blake2b256)
+    #[serde(serialize_with = "as_base64", deserialize_with = "from_base64")]
+    pub proof: Vec<u8>,
+    #[serde(skip_serializing_if = "String::is_empty", default = "Default::default")]
+    pub comment: String,
+}
+
 /// A `Common` part of every `Content` format
 #[derive(Clone, Builder, Debug, Serialize, Deserialize)]
 pub struct Common {
@@ -70,6 +80,9 @@ pub struct Common {
     pub date: chrono::DateTime<FixedOffset>,
     /// Author of the proof
     pub from: crate::PublicId,
+    /// Reference to original proof when reissuing
+    #[serde(skip_serializing_if = "Option::is_none", default = "Option::default")]
+    pub original: Option<OriginalReference>,
 }
 
 impl CommonOps for Common {
@@ -150,10 +163,12 @@ pub struct Draft {
 }
 
 impl Draft {
+    #[must_use]
     pub fn title(&self) -> &str {
         &self.title
     }
 
+    #[must_use]
     pub fn body(&self) -> &str {
         &self.body
     }
@@ -197,7 +212,7 @@ pub trait ContentExt: Content {
             digest: crev_common::blake2b256sum(body.as_bytes()),
             body,
             signature: crev_common::base64_encode(&signature),
-            common_content: self.common().to_owned(),
+            common_content: self.common().clone(),
         };
         let parsed = proof::Proof::parse_from(std::io::Cursor::new(proof.to_string().as_bytes()))?;
 
